requirements.txt
PySide6
watchdog
pillow
ursina
